+++ src/textscreens.cpp
@@ -304,11 +304,14 @@ void TextWindow::ShowGroupInfo() {
     if(g->type == Group::Type::LATHE) {
         Printf(true, " %Ftlathe plane sketch");
     } else if(g->type == Group::Type::EXTRUDE || g->type == Group::Type::ROTATE ||
-              g->type == Group::Type::TRANSLATE || g->type == Group::Type::REVOLVE) {
+              g->type == Group::Type::TRANSLATE || g->type == Group::Type::REVOLVE ||
+              g->type == Group::Type::HELIX) {
         if(g->type == Group::Type::EXTRUDE) {
             s = "extrude plane sketch";
         } else if(g->type == Group::Type::TRANSLATE) {
             s = "translate original sketch";
+        } else if(g->type == Group::Type::HELIX) {
+            s = "create helical extrusion";
         } else if(g->type == Group::Type::ROTATE) {
             s = "rotate original sketch";
         } else if(g->type == Group::Type::REVOLVE) {
@@ -364,7 +367,8 @@ void TextWindow::ShowGroupInfo() {
     Printf(false, "");
 
     if(g->type == Group::Type::EXTRUDE || g->type == Group::Type::LATHE ||
-       g->type == Group::Type::REVOLVE || g->type == Group::Type::LINKED) {
+       g->type == Group::Type::REVOLVE || g->type == Group::Type::LINKED ||
+       g->type == Group::Type::HELIX) {
         bool un   = (g->meshCombine == Group::CombineAs::UNION);
         bool diff = (g->meshCombine == Group::CombineAs::DIFFERENCE);
         bool asy  = (g->meshCombine == Group::CombineAs::ASSEMBLE);
@@ -384,7 +388,7 @@ void TextWindow::ShowGroupInfo() {
             (asy ? RADIO_TRUE : RADIO_FALSE));
 
         if(g->type == Group::Type::EXTRUDE || g->type == Group::Type::LATHE ||
-           g->type == Group::Type::REVOLVE) {
+           g->type == Group::Type::REVOLVE || g->type == Group::Type::HELIX) {
             Printf(false,
                 "%Bd   %Ftcolor   %E%Bz  %Bd (%@, %@, %@) %f%D%Lf%Fl[change]%E",
                 &g->color,
@@ -396,7 +400,8 @@ void TextWindow::ShowGroupInfo() {
         }
 
         if(g->type == Group::Type::EXTRUDE || g->type == Group::Type::LATHE ||
-           g->type == Group::Type::REVOLVE || g->type == Group::Type::LINKED) {
+           g->type == Group::Type::REVOLVE || g->type == Group::Type::LINKED ||
+           g->type == Group::Type::HELIX) {
             Printf(false, "   %Fd%f%LP%s  suppress this group's solid model",
                 &TextWindow::ScreenChangeGroupOption,
                 g->suppress ? CHECK_TRUE : CHECK_FALSE);