+++ src/ui.h
@@ -125,6 +125,7 @@ enum class Command : uint32_t {
     GROUP_3D,
     GROUP_WRKPL,
     GROUP_EXTRUDE,
+    GROUP_HELIX,
     GROUP_LATHE,
     GROUP_REVOLVE,
     GROUP_ROT,